+++ react/src/features/game/talk/hooks/useTalkData.ts
@@ -1,17 +1,23 @@
 import { useEffect, useState } from 'react';
 import { type CoRole } from '../../type';
 import { fetchTalkIndex } from '../api';
+import { UnexpectedError } from '@/features/error';
+import { useGameIndex } from '@/features/game/hooks/useGameIndex';
+import type { Player } from '@/features/game/talk/type';
 
 export const useTalkData = (): {
   gameId: number | undefined;
   nightActLog: string | undefined;
   hostFlg: boolean | undefined;
-  cos: CoRole[] | undefined;
+  players: Player[] | undefined;
+  setPlayers: React.Dispatch<React.SetStateAction<Player[] | undefined>>;
+  getMyPlayer: () => Player;
 } => {
   const [gameId, setGameId] = useState<number | undefined>(undefined);
   const [nightActLog, setNightActLog] = useState<string | undefined>(undefined);
   const [hostFlg, setHostFlg] = useState<boolean | undefined>(undefined);
   const [cos, setCos] = useState<CoRole[] | undefined>(undefined);
+  const [players, setPlayers] = useState<Player[] | undefined>(undefined);
 
   useEffect(() => {
     const fetchTalkIndexAsync = async () => {
@@ -25,10 +31,52 @@ export const useTalkData = (): {
     void fetchTalkIndexAsync();
   }, []);
 
+  const { playerId, playerName, playerRole, otherPlayerList } = useGameIndex(
+    'talk',
+    gameId,
+  );
+
+  useEffect(() => {
+    if (
+      playerId === undefined ||
+      playerName === undefined ||
+      playerRole === undefined ||
+      otherPlayerList === undefined ||
+      cos === undefined
+    )
+      return;
+
+    const players: Player[] = [
+      {
+        id: playerId,
+        name: playerName,
+        role: playerRole,
+        co: cos.find((co) => co.id === playerId)!,
+      },
+      ...otherPlayerList.map((otherPlayer) => ({
+        id: otherPlayer.id,
+        name: otherPlayer.name,
+        role: otherPlayer.role,
+        co: cos.find((co) => co.id === otherPlayer.id)!,
+      })),
+    ];
+    setPlayers(players);
+  }, [playerId, playerName, playerRole, otherPlayerList, cos]);
+
+  const getMyPlayer = (): Player => {
+    if (!players) {
+      throw new UnexpectedError('players is undefined');
+    }
+
+    return players[0];
+  };
+
   return {
     gameId,
     nightActLog,
     hostFlg,
-    cos,
+    players,
+    setPlayers,
+    getMyPlayer,
   };
 };
